+++ scripts/get_RGBD-S_from_a_scene.py
@@ -196,9 +196,24 @@ def main(dataset):
             node_list_to_travel = my_graph.nodes()
             angle_list_to_travel = []
 
-        # Assign the node-angle anchor to the first node in the node_list_to_travel
-        anchor_node_angle = f"{node_list_to_travel[0]}_{0 if len(angle_list_to_travel) == 0 else angle_list_to_travel[0]}"
-        data_writer.add_anchor_of_relative_transforms(anchor_node_angle) # Add anchor of relative transforms
+        # Store the pose of all nodes in the scene's graph to a file.
+        for n, node in enumerate(my_graph.nodes()):
+            position = my_graph.nodes[node]["point"] # (3,)
+            for angle in range(0, 360, 10):
+                image_id = f"nodeID_{node}_angleID_{angle}"
+                rotation = quat_to_coeffs(
+                    quat_from_angle_axis(np.deg2rad(angle), np.array([0, 1, 0]))
+                ).tolist()  # [b, c, d, a] where the unit quaternion would be a + bi + cj + dk
+                # Transformation from the agent's camera frame to the world frame
+                cvCam2W_T  = simulator.get_opencvCam_to_world_transformation(
+                    position,
+                    [rotation[-1]] + rotation[:-1], # [b, c, d, a] -> [a, b, c, d]
+                )
+                # We actually want to save the transformation from the camera frame to the world frame. However,
+                # the rotation between the camera frame and the world frame is as same as the rotation between the agent frame and the world frame.
+                data_writer.add_pose_in_metrics_space(image_id, cvCam2W_T)
+
+
         for n, node in enumerate(node_list_to_travel):
             print(f"-------------------------------------------")
             # All rotation and position here are w.r.t the habitat's coordinate system which is
@@ -250,15 +265,6 @@ def main(dataset):
                     rgbd_sstate.position,
                     [rgbd_sstate.rotation.w] + list(rgbd_sstate.rotation.vec),
                 )
-
-                # Take one node as the anchor and find the relative transformation between any node to that node. This will be useful
-                # later on in sound NeRF to find each node's coordinate in the NeRF space.
-                if f"{node}_{angle}"== anchor_node_angle:
-                    anchor_cvCam2W_T = obs["cvCam2W_T"].copy()
-                    print(f"[Info] This node {node} angle {angle} is used as the anchor for other poses")
-                    print(f"[Info] anchor_cvCam2W_T: {anchor_cvCam2W_T}")
-                cvCam2anchor_T = np.linalg.inv(anchor_cvCam2W_T) @ obs["cvCam2W_T"]
-                print(f"[Info] cvCam2anchor_T: {cvCam2anchor_T}")
 
                 # Convert the observation to an image frame for demo videos
                 frame = convert_observation_to_frame(obs)
@@ -344,11 +350,9 @@ def main(dataset):
 
                 data_writer.add_colmap_image(image_id, colmap_image_instance)
 
-                # Write the relative transformation from the camera to the anchor node - angle
-                data_writer.add_relative_transform_to_anchor(
-                    image_id = image_id,
-                    cvCam2anchor_T =cvCam2anchor_T,
-                )
+                # # Write the pose in the metrics space
+                # data_writer.add_pose_in_metrics_space(image_id = image_id, cvCam2world_T=obs["cvCam2W_T"])
+
                 # RGBD-S data
                 if "rgb" in args.modalities.lower():
                     data_writer.add_rgb_image(f"{image_id}.jpg", obs)
@@ -371,9 +375,9 @@ def main(dataset):
         data_writer.write_colmap_data_to_files(".txt")
         print(f"[Info] Saved colmap data from scene {args.scene} to\n {scene_obs_dir}")
 
-        # Write relative transformation from the camera to the anchor node - angle
-        data_writer.write_transform2anchor_to_files(".json")
-        print(f"[Info] Saved transformations from camera poses to the anchor to\n {scene_obs_dir}")
+        # Write the poses in the metrics space to file
+        data_writer.write_poses_in_metrics_space_to_file(".json")
+        print(f"[Info] Saved poses in metrics space to\n {scene_obs_dir}")
 
         # Write RGB-S data
         data_writer.write_rgbds_data_to_files()